refactor: Amenities, Room 파일분리
