+++ rooms/urls.py
@@ -1,15 +1,21 @@
 from django.urls import path
-from . import views
-from rooms.views import RoomsListAPI, RoomCreateAPI
-
+from rooms.views.view_v0_amenity import Amenities, AmenityDetail
+from rooms.views.view_v1_room import (
+    RoomsListAPI,
+    RoomCreateAPI,
+    RoomDetail,
+    RoomReviews,
+    RoomAmenities,
+    RoomPhotos,
+)
 
 urlpatterns = [
-    path("v0/list", RoomsListAPI.as_view()),
+    path("v1/list", RoomsListAPI.as_view()),
     path("v1/create", RoomCreateAPI.as_view()),
-    path("v0/<int:pk>", views.RoomDetail.as_view()),
-    path("v0/<int:pk>/reviews", views.RoomReviews.as_view()),
-    path("v0/<int:pk>/amenities", views.RoomAmenities.as_view()),
-    path("v0/<int:pk>/photos", views.RoomPhotos.as_view()),
-    path("v0/amenities/all", views.Amenities.as_view()),
-    path("v0/amenities/<int:pk>", views.AmenityDetail.as_view()),
+    path("v0/<int:pk>", RoomDetail.as_view()),
+    path("v0/<int:pk>/reviews", RoomReviews.as_view()),
+    path("v0/<int:pk>/amenities", RoomAmenities.as_view()),
+    path("v0/<int:pk>/photos", RoomPhotos.as_view()),
+    path("v0/amenities/all", Amenities.as_view()),
+    path("v0/amenities/<int:pk>", AmenityDetail.as_view()),
 ]
