+++ rooms/views/view_v0_amenity.py
@@ -1,0 +1,58 @@
+from rest_framework.views import APIView
+from rest_framework.status import HTTP_204_NO_CONTENT
+from rest_framework.response import Response
+from rest_framework.exceptions import (
+    NotFound,
+)
+from rooms.models import Amenity
+from rooms import serializers
+
+
+class Amenities(APIView):
+    def get(self, request):
+        all_amenities = Amenity.objects.all()
+        serializer = serializers.AmenitySerializer(all_amenities, many=True)
+        return Response(serializer.data)
+
+    def post(self, request):
+        serializer = serializers.AmenitySerializer(data=request.data)
+        if serializer.is_valid():
+            amenity = serializer.save()
+            return Response(
+                serializers.AmenitySerializer(amenity).data,
+            )
+        else:
+            return Response(serializer.errors)
+
+
+class AmenityDetail(APIView):
+    def get_object(self, pk):
+        try:
+            return Amenity.objects.get(pk=pk)
+        except Amenity.DoesNotExist:
+            raise NotFound
+
+    def get(self, request, pk):
+        amenity = self.get_object(pk)
+        serializer = serializers.AmenitySerializer(amenity)
+        return Response(serializer.data)
+
+    def put(self, request, pk):
+        amenity = self.get_object(pk)
+        serializer = serializers.AmenitySerializer(
+            amenity,
+            data=request.data,
+            partial=True,
+        )
+        if serializer.is_valid():
+            updated_amenity = serializer.save()
+            return Response(
+                serializers.AmenitySerializer(updated_amenity).data,
+            )
+        else:
+            return Response(serializer.errors)
+
+    def delete(self, request, pk):
+        amenity = self.get_object(pk)
+        amenity.delete()
+        return Response(status=HTTP_204_NO_CONTENT)
